Use `tokenizer.decode` to retrieve token string
We currently have non-digit tokens in the sequences output by
the models when `type` is "int" or "float".

Somehow we do not get the same character when iterating over the
vocabulary and when using `tokenizer.decode`; since the latter
is used to decode the final sentence we use this when creating
the mask.