Use FSM-based scanning
All the terminal symbols regexs in each parse-state-dependent lexer are
combined/unioned into a single FSM, and scanning is performed according to those
combined FSMs.  The function `fsm_union` was added for that purpose.

Since the parser needs to know exactly which terminal symbols were matched, we
now need to the ability to determine exactly which sub-FSM (i.e. one of the
combined terminal symbol FSMs) accepted an input string.  The function
`get_sub_fsms_from_seq` serves this purpose.