Convert tokens to strings before partial matching
BPE tokenizers encode whitespaces as a special character, and it is
possible that another tokenizer may encode one or several strings as
special character(s).

We thus add a `convert_token_to_string` that converts back these special
character to the string they correspond to so we can partially match
the corresponding tokens with the regex.