Improve multiple-choice selection for the OpenAI API
The current approach is greedy, in the sense that it generates a single
token at each steps, asking the API to only generate valid next tokens.
This mean having to pay for the prompt tokens for every token generated.

This commit takes a more optimistic approach. It starts with allowing
all tokens present in the sequences, and limiting the length of the
generation to the number of tokens in the longest sequence. If the
completion is not satisfactory it then takes one greedy step before
switching back to the optimistic mode.

On average this new approach consumes less tokens than the current one.