Added subclass of OpenAI class to support custom endpoints (#560)
Adds a new class to support for providing a custom base_url to the
OpenAI client, and a couple of small updates to remove the checks on the
model via the "retrieve" api (since it's not always available), sets
tiktoken to use gpt4 by default (with the option to override it), and
enables passing temperature as a named parameter to the __call__method
to override the config.

Usage is basically the same as the regular OpenAI model wrapper, with
the addition of the base_url parameter. Eg:

```
hermes = CustomOpenAI(
        model_name="NousResearch/Nous-Hermes-2-Mixtral-8x7B-DPO",
        system_prompt=system_prompt,
        api_key="xxx",
        base_url='https://api.together.xyz',
 )

print(hermes(test_input))

```
The environment variables checked to set the API key was also updated,
so as not to conflict with any OpenAI configurations in place, and sound
a little more generic, along with a new env var `INFERENCE_BASE_URL` to
act as an alternative to the constructor parameter.