Update modal example (#1111)
I updated some of the text of the modal example to add more information
on what's happening where, since Modal's general structure may be
unfamiliar to new users.

List of changes:

- Removed version constraints on the image. I think this is a
questionable choice so I'd love input -- in general, this should
probably better be addressed by versioning the docs #999. There's an
open PR #1059 here that can help with this too -- I'm happy to revert to
pinned versions, or at least update to the current version of outlines
(0.0.46).
- Added a brief note that the Mistral model is gated and requires you to
request access on HuggingFace.
- Moved the environment variable setting from `import_model` to a `.env`
call when the image is created. This is more idiomatic Modal code. It
was previously addressed in #1058, but this should be more current.
- More comments and prep.