Fix broken links in the README (#1625)
On top of the missing links highlighted in #1622, there were a bunch of
broken links in the README. This PR makes sure all links lead to
actually existing page. In some cases, the links do not lead to the
exact section it should, but that's a problem to fix in the
documentation in a separate PR. In the meantime, links at least lead to
a related page from which the user should be able to find their way.